+++ CHANGELOG.md
@@ -6,6 +6,14 @@ View all releases at: https://github.com/trimble-oss/modus-web-components/releas
 
 ## Unreleased
 
+### Added
+
+- Added small Modus Chip variant
+
+### Removed
+
+- Removed large Modus Chip variant
+
 ## 0.1.12 - 2022-22-08
 
 ### Added